libcrun: remove unused intprops.h
This was imported from gnulib with commit
6c36eb46ca6ef678a7994265cbe69afcae4cd567 to implement sig2str. However,
sig2str was replaced with 1f94ee6f25aff44292428fd9d9baaeb8f6eddbd2, and
made this header file redundant.

The header itself is problematic from a licensing point of view due to a
bit of a checkered history: the file states that it's copyrighted by the
FSF, but licensed under the GPL (not LGPL) despite being part of
libcrun. Additionally, the stated license is GPL v2.1 (which does not
exist); this was likely a typo, during a mass GPL->LGPL relicensing that
occured with commit 34a0ff3cc9fc588bf27d959987b1ecbea6261e15.

The original file from gnulib d41bb6138ba9cfcc9445a9ba938f1361defc2576
was licensed under the LGPLv2.1+, and imported into this repository as
GPLv3 at the time. Beyond licensing header changes and clang-format, it
did not receive any other changes, nor any code was derived from it.

All in all, no historical viral GPL influences are present, and the file
can be safely removed with no ill effects, as it remains unused.

Signed-off-by: Faidon Liambotis <paravoid@debian.org>