libcrun: add alias argument to function
now write_file_and_check_controllers_at accepts a name alias argument,
if the first file is not found and the operation fails with ENOENT,
then it tries to open the file with the alias name.  It is useful to
simplify some existing code and avoid the same logic in multiple
places.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>