container: add custom annotation to specify the scheduler
This commit adds a new feature to the container runtime that allows
users to set the scheduling policy of the container process.

The new feature is implemented as a custom annotation.  To set the
scheduling policy and priority, users can add a `run.oci.scheduler`
annotation with a value in the format POLICY[|OPTION][#PRIORITY].

If no scheduling policy or priority is specified, the container process
will use the current scheduling policy and priority.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>