wasm, wasmedge: add current directory to preopen paths
A similar feature was added to crun `wasmtime` (
https://github.com/containers/crun/pull/985 ) where we decided to
`preopen` current directory for all the instances, this PR attempts same
for `wasmedge`.

There is no generic external method so users can pass this to handler as
discussed here: https://github.com/containers/crun/pull/985 but there is
no need shown till now so idea is to proceed `.` as a default directory
which whill be pre-opened.

Signed-off-by: Aditya R <arajan@redhat.com>