cgroup: workaround cpu quota/period issue with v1
Sometimes setting CPU quota period fails when a new period is lower,
and a parent cgroup has CPU quota limit set.

This happens as in cgroup v1 the quota and the period can not be set
together (this is fixed in v2), and since the period is being set first,
new_limit = old_quota/new_period may be higher than the parent cgroup
limit.

The fix is to retry setting the period after the quota, to cover all
possible scenarios.

Tested via runc integration tests. Before the commit, it fails:

	root@ubu2004:~/git/runc# RUNC=`pwd`/../crun/crun.before bats -f "pod cgroup" -t  tests/integration/update.bats
	1..1
	not ok 1 update cpu period in a pod cgroup with pod limit set
	# (in test file tests/integration/update.bats, line 424)
	#   `[ "$status" -eq 0 ]' failed
	# crun.before spec (status=0):
	#
	# crun.before run -d --console-socket /tmp/bats-run-30428-dYkMDC/runc.4FdCtn/tty/sock test_update (status=0):
	#
	# crun.before update --cpu-quota 600000 test_update (status=1):
	# writing file `cpu.cfs_quota_us`: Invalid argument
	# crun.before update --cpu-period 10000 --cpu-quota 3000 test_update (status=1):
	# writing file `cpu.cfs_period_us`: Invalid argument

With the fix, the test passes.

Originally reported for runc in
https://github.com/opencontainers/runc/issues/3084

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>