tests: fix cpu-weight-systemd test under a user manager
The crun systemd cgroup manager tries to connect to the user manager
first, and only falls back to the system manager if that fails
(open_sd_bus_connection in src/libcrun/cgroup-systemd.c)

Thus the scope is created under user.slice -> user-0.slice ->
user@0.service -> app.slice -> crun-test-tmp4_wkotzy.scope.

The test tries to validate that CPUWeight was set by running (e.g.):
  systemctl show crun-test-tmp4_wkotzy.scope
This fails, because --system is implied and there is no such scope in
the system manager, as the scope was created in the user manager.
The correct command in this case would be:
  systemctl --user show crun-test-tmp4_wkotzy.scope

This test failure was not caught up until now because a) the test is
skipped if rootless, b) root usually is not logged in into a tty, and a
user manager (DBus session) does not exist. However it /can/ happen, and
did happen in the Debian autopkgtest environment.

Fix this by trying "systemctl show" twice: once against the system
manager and once against the user manager. The fix is kinda ugly and
repetitive, but abstracting this for two callsites looked even uglier.

Closes: https://github.com/containers/crun/issues/1197

Signed-off-by: Faidon Liambotis <paravoid@debian.org>