linux: add sync before sending mounts
add a sync before sending the mounts to the container.  This is needed
when the container needs to join an existing PID namespace and this is
done later once the container PID is already created.  Without the
sync, we could hit a race condition where the container namespaces
were not joined yet and the /proc/$PID/ns/user would still refer to
the old user namespace.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>