libcrun: handle SIGWINCH by resizing terminal_fd
When receiving the WINCH signal, crun was using the default signal
behavior of passing it through to the child process. A child that cared
about the terminal size would then call the TIOCGWINSZ ioctl, but find
the size had not changed. This is due to crun not copying the terminal
size for its stdin to the child's pty.

This changeset handles the WINCH signal by getting crun's terminal size,
and setting the child's pty to match. The WINCH signal is not passed
through, as the kernel will send it as part of handling the TIOCSWINSZ
ioctl.

Signed-off-by: Terin Stock <terinjokes@gmail.com>