linux: simplify setns with pidfd
This caused an issue when the PID 1 didn't execve yet the container
process, and joining the user namespace first would cause the user to
look access to the container PID 1 namespaces.

When running with enough permissions, setns() will allow to join all
the namespaces in one shoot, so let's use it.

On any error, attempt the fallback mechanism to join each namespace
separately.

Closes: https://github.com/containers/crun/issues/1264

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>