features: Fix annotations formatting
The runtime spec expects strings as values and keys in the annotations
sections. This is said here[1]:

	Annotations MUST be a key-value map that follows the same
	convention as the Key and Values of the annotations property of
	config.json

And the annotations docs say:

	Values MUST be strings. Values MAY be an empty string.

This patch just writes the string "true" when the bool is true and
"false" otherwise.

I noticed this while trying to parse the output of "crun features" with
https://github.com/containerd/go-runc, that parses the json to the
runtime-spec provided go-bindings that expect a string, and fail.
Without this patch, parsing the output fails with:

	cannot unmarshal bool into Go struct field Features.annotations of type string: unknown

[1]: https://github.com/opencontainers/runtime-spec/blob/e8c413418efcb816b963be9762ddc202e181ba88/features.md#annotations
Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>