features: Support mountExtensions
This PR upstream added the spec for the mountExtensions feature field:
	https://github.com/opencontainers/runtime-spec/pull/1219

This commit just implements that and updates the OCI max version
implemented to the one currently used in the spec (an unreleased
version).

It is not clear if in the future, the version of unreleased specs will
be changed to something else:
	https://github.com/opencontainers/runtime-spec/pull/1221

But this is what is currently accepted.

Signed-off-by: Rodrigo Campos <rodrigoca@microsoft.com>