utils: improve error message for ensure_dir
when creating a directory, if any component of the path is not a
directory, report it as part of the error message.

Forcing O_DIRECTORY for any component that we expect to be a
directory, will force the kernel to report the error earlier instead
of failing at the next mkdir(2) call.

Closes: https://github.com/containers/crun/issues/1307

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>