linux: open mounts before setgroups if in a userns
extend the condition for opening bind mounts in the parent process to
the case where the crun process itself is running inside a user
namespace.  In this way the source path is opened before doing the
setgroups(2) syscall and potentially losing access to directories that
are accessible only by group permission.

Closes: https://github.com/containers/crun/issues/1320

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>