linux: new mount option "copy-symlink"
provide an option to copy the original symlink instead of
dereferencing it.  This requires a new mount flag, which is handled
directly by crun and not passed down to the kernel.  When "copy-symlink"
is found, if the source of a mount is a symlink, then crun creates the
symlink at the destination instead of attempting a mount(2).

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>