linux: append tmpfs mode if missing for mounts
For tmpfs type mounts, ensure that a mode is appended to the mount
data if it is not already present.  The function checks the
existing directory and appends the mode based on the file permissions
of the source path at the rootfs.  This helps in preserving the
intended file permissions when overriding a directory with a new tmpfs
file system.

Closes: https://issues.redhat.com/browse/RHEL-14469

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>