Packit: reuse non-RHEL failure message notification on RHEL
Packit posts an additional failure notification comment if the latest
failure message does not match the previous notification comment.

This commit reuses the non-RHEL failure notification comment on RHEL
tests. This should reduce the number of failure messages posted on a PR.

For reference, see the alternating failure messages on #1425 for RHEL
and non-RHEL.

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>