status: rmdirfd: try harder to remove mount points
The previous patch left a mount point in /run/crun which could not be
removed, we should try harder to remove directories failing with EBUSY
as we do in create_missing_devs().

This patch also fixes the issue described in the previous commit
("status: rmdirfd: try harder to remove mount points").

Fixes: 523eed3f0f4b ("linux: add new fallback when mount fails with EBUSY")
Signed-off-by: Dominique Martinet <dominique.martinet@atmark-techno.com>