Fix running on kernel without user namespaces
As long as container is not rootless, runc can run it on a Kernel
without enabled user namespaces, which makes sense to me. I think crun
should be able to do the same, however I hit some issues with it.

First I got
    error opening file `/proc/self/uid_map`: No such file or directory
and that's because when kernel doesn't have user namespaces,
the uid_map doesn't exist at all. Adding a check for file existence,
after the read_all_file call, solves the issue.

With that change in place I got an error about next missing file.
This time it was /proc/self/setgroups and it was coming from
can_setgroups call. Adding similar fix to the one for uid_map, solves
the issue.

With both changes in place I am now able to run a non-rootless
container on a kernel without user namespaces enabled.

Signed-off-by: Michal Sieron <michalwsieron@gmail.com>