status: validate container id
check that the container name does not contain any slash so that
operations in status.c won't risk to access directories/files outside
of the run directory.

Beside, it is not a security issue because if a directory exists, crun
refuses to use it; and if it does not exist then it is created and
immediately deleted because the container creation fails.

Play safe and do not allow any slash in the container name.

Closes: https://github.com/containers/crun/issues/1646

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>