krun: implement support for external kernels
libkrun has recently acquired the ability to load external kernels. Use
it to enable users to bundle kernel images in their container files.

Note: in our execution model, the kernel in the microVM is not (and has
never been) considered a trusted component. This is the reason why we
put the whole VMM inside the container and avoid using vhost or any
other features that require special privileges, beyond KVM.

Signed-off-by: Sergio Lopez <slp@redhat.com>