crun: Fix the binary name in the Usage info
This patch solves the bug where the Usage info for
commands displays the command name instead of "crun"
as the crun binary name.
E.g. for `crun list --help` crun prints "Usage: list [OPTION...] list"
while it should print "Usage: crun [OPTION...] list".

This fix clones and modifies the `argc` and `argv`
for the second `argp_parse` execution - the `parse_opt`
of each command will now receive the binary name (crun)
instead of the command name in argv[0].
The change is transparent to commands, because
- the command name is extracted and used in the first
`argp_parse` execution in crun, and
- the argp_parse doesn't call the parse_opt with the argv[0]
by default

Signed-off-by: Piotr Romanski <piotrek.romanski@gmail.com>