criu: allow mount type to be NULL
In runtime-spec, mount's type property is optional and thus it can be
NULL (for bind mounts). This is accounted for in other parts of code
but not in c/r.

This fix prevents SIGSEGV on crun checkpoint/restore.

Found when running runc's integration test named
"checkpoint and restore (bind mount, destination is symlink)"

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>