linux: use rootfsfd directly from container data
This change ensures that the file descriptor for the
rootfs is always sourced directly from the container's
private data.  This avoids potential stale file descriptor
issues that could happen if a local variable were used and
the descriptor in the private data was updated elsewhere.

Should not introduce any behavior change.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>