Fixup "criu: support mounts where dest is a symlink"
Commit e0b01580 "criu: support mounts where dest is a symlink" brought
in a few issues.

First, the concatenation of status->bundle and status->rootfs do not
make any sense (and I'm not yet sure why it is used as CRIU root).
Revert to using status->rootfs directly.

Second, if chroot_resolve returns ENOENT, it is OK (not all paths are
visible from the host -- some are only inside the mount namespace).
In this case, use the destination as is.

This should fix podman checkpoint failures.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>