linux: add `nofollow` option to `get_bind_mount`
This change enhances the `get_bind_mount` function by
introducing a `nofollow` boolean parameter.  When set to
true, this parameter ensures that the `AT_SYMLINK_NOFOLLOW`
flag is passed to the `syscall_open_tree` call.

This allows the caller to specify that if the source of
the bind mount is a symbolic link, the link itself should
be mounted, rather than the target it points to.

All existing call sites have been updated to pass `false`
for the new `nofollow` parameter, thus preserving the
previous behavior of following symlinks by default. This
addition provides finer control for future use cases
requiring specific handling of symbolic links during
bind mount creation.

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>