cgroup: do not create a sub-cgroup by default
The `run.oci.systemd.subgroup` annotation previously defaulted
to creating a "container" sub-cgroup on cgroup v2 systems if
the annotation was not specified.  This change alters the
behavior so that no sub-cgroup is created by default.

A sub-cgroup will only be created if the
`run.oci.systemd.subgroup` annotation is explicitly provided
with a non-empty value.

Closes: https://github.com/containers/crun/issues/1766

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>