cgroup,systemd: allow empty slice in cgroupsPath
While it may not be properly documented in runtime-spec, runc's systemd
cgroup driver follows some rules when constructing the slice and scope
from the linux.cgroupsPath (see [1]).

One such rule is, when the slice is empty, it defaults to "system.slice",
unless we have cgroup v2 and a rootless container, in which case it
defaults to "user.slice". This is supported by runc and although it
might be questionable, it makes sense for crun to be compatible.

Add a test case.

Fixes: 1811.

[1]: https://github.com/opencontainers/runc/blob/main/docs/systemd.md#systemd-unit-name-and-placement

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>