cgroup,systemd: check for loaded bpf on update
Since systemd does not support updating BPFProgram property, we can't
really perform updates to device access list. Yet, there is a common
scenario (with podman) when crun update receives the same set of devices
as during the container creation. In such case, we can obtain the
old/original bpf program and compare it with the new one. If they match,
we can continue without doing anything.

If they don't, we still have to error out.

[@kolyshkin: commit message]

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>