container: propagate error from exec setup
commit 7b82568080dfd79f3bf38c47f82e38d4878fb2c9 caused a double error
message to be printed when the exec session setup fails.

It is not strictly a bug, since there is nothing wrong to show an
additional error line but that caused Podman to be confused and not
handle the error string correctly.

Partially revert the change, now errors are propagated to the parent
process so that a single error line is printed.

Closes: https://github.com/containers/crun/issues/1823

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>