krun: avoid failing if sev/nitro are not available
It's possible for a user to accidentally install the sev or nitro
flavors of libkrun even if their system doesn't actually support those
features.

Before this change, having them installed on a system that doesn't
support them would cause krun to stop working. Let's do the right
thing and, if the bindings for "/dev/sev" or "/dev/nitro_enclaves"
can't be created, close the handles to those library flavors and keep
running with the generic one.

If sev or nitro where really needed by the workload, it'll fail
gracefully in "libkrun_configure_flavor".

Fixes: #1120 #1300

Signed-off-by: Sergio Lopez <slp@redhat.com>