cgroup, systemd: improved "same bpf" check
It appears that some architectures other than x86_64 (ppc64 and s390) do
modify in-kernel bpf program while it's being loaded into the kernel.
On those platforms, the check for bpf being the same (added in commit
edf667889) doesn't work, and the attempt to run "crun update" with
the same device access list fails with "updating device access list not
supported when using BPFProgram" error.

A workaround is to actually load the new program into the kernel, read
it back, and compare.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>