criu: enable setting of RPC config file
This commit adds support for specifying a CRIU RPC configuration file.
This config file allows users to overwrite the default CRIU options
used by the container runtime, for example, to specify options such as
`--tcp-established` or `--tcp-close` when checkpointing containers
with TCP connections in Kubernetes.

For compatibility with runc, the default config file path is set to
`/etc/criu/runc.conf`. We also introduce support for crun.conf
that will be used instead of runc.conf when the file is available.

`criu_set_config_file()` was added to libcriu in version 4.2

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>